Prepare for Blackwood chapter
The quests used for determining if skill points from the main quest
were earned are nothing more than an educated guess.

As I did not play the story on the PTS (nor do I intend to) this has to
be adjusted afterwards. The skyshards and public dungeons
should already work though.

An override for a skill point of new Blackwood chars is not
implemented either. This may have to be added if need be.